dear christ i'm frustrated
I'm trying to become more fluent in this stuff, but...holy crap, none of this is coming easily. I still barely understand why we're using GitHub, nor really how to use it. I want to comment on previous files in my repos, but how could I do that? This is more a rant than a meaningful commit, so I'll change pace in a sec, but I seriously need help.

But anyway...in terms of a meaningful commit message, I've changed many of the fonts around, messed with the image placement, and made the whole thing as clean as possible. Satisfying. However, I forgot to read the section about the two versions of .png files. That was lost on me because of the frustration I experienced with working layers, scaling stuff, and generally trying to get the text to align at all. Which is there a more efficient way to move text boxes than clicking specifically on the letters in the box?